refactor(wallet): split implementation into feature modules
Follow the market/ layout precedent. Move the 2253-line
WalletPlugin.cpp into wallet/ with per-feature files:

- WalletPlugin.cpp: lifecycle, events, transfer, config
- WalletGui.cpp: command and GUI registration
- WalletRedEnvelope.cpp: send, grab, refund, sweep, stats
- WalletBank.cpp: bank deposits and wealth ranking
- WalletLedger.cpp: ledger, fee pool, daily stats

Shared includes, helper and struct definitions live in
include/server/Plugins/wallet/WalletDetail.h (MarketGuiDetail.h
pattern). Pure code movement, no behavior change.
